Unison.Util.BiMultimap

Description

A left-unique relation.

Documentation

data BiMultimap a b

A left-unique relation.

"Left-unique" means that for all (x, y) in the relation, y is related only to x.

empty :: (Ord a, Ord b) => BiMultimap a b

An empty left-unique relation.

Basic queries

isEmpty :: BiMultimap a b -> Bool

Is a left-unique relation empty?

Lookup

memberDom :: Ord a => a -> BiMultimap a b -> Bool

lookupDom :: Ord a => a -> BiMultimap a b -> Set b

Look up the set of b related to an a.

O(log a).

lookupRan :: Ord b => b -> BiMultimap a b -> Maybe a

Look up the a related to a b.

O(log b).

unsafeLookupRan :: Ord b => b -> BiMultimap a b -> a

Look up the a related to a b.

O(log b).

lookupPreimage :: (Ord a, Ord b) => b -> BiMultimap a b -> Set b

Look up the preimage of a b, that is, the set of b that are related to the same a as the input b.

/O(log a + log b)

Mapping / traversing

unsafeTraverseDom :: forall a b m x. (Monad m, Ord b, Ord x) => (a -> m b) -> BiMultimap a x -> m (BiMultimap b x)

Traverse over the domain a left-unique relation.

The caller is responsible for maintaining left-uniqueness.

Filtering

filter :: (Ord a, Ord b) => (a -> b -> Bool) -> BiMultimap a b -> BiMultimap a b

Filter a left-unique relation, keeping only members (a, b) that satisfy a predicate.

filterDom :: (Ord a, Ord b) => (a -> Bool) -> BiMultimap a b -> BiMultimap a b

Filter a left-unique relation, keeping only members (a, b) whose a satisfies a predicate.

filterDomain :: (Ord a, Ord b) => (a -> NESet b -> Bool) -> BiMultimap a b -> BiMultimap a b

Filter a left-unique relation, keeping only members (a, b) whose a and set of b satisfies a predicate.

restrictDom :: (Ord a, Ord b) => Set a -> BiMultimap a b -> BiMultimap a b

Restrict a left-unique relation to only those (a, b) members whose a is in the given set.

restrictRan :: (Ord a, Ord b) => Set b -> BiMultimap a b -> BiMultimap a b

Restrict a left-unique relation to only those (a, b) members whose b is in the given set.

withoutDom :: (Ord a, Ord b) => Set a -> BiMultimap a b -> BiMultimap a b

Restrict a left-unique relation to only those (a, b) members whose a is not in the given set.

withoutRan :: (Ord a, Ord b) => Set b -> BiMultimap a b -> BiMultimap a b

Restrict a left-unique relation to only those (a, b) members whose b is not in the given set.

Maps

domain :: BiMultimap a b -> Map a (NESet b)

range :: BiMultimap a b -> Map b a

O(1).

unsafeFromDomain :: Ord b => Map a (NESet b) -> BiMultimap a b

Construct a left-unique relation from a mapping from its left-elements to set-of-right-elements. The caller is responsible for ensuring that no right-element is mapped to by two different left-elements.

fromRange :: (Ord a, Ord b) => Map b a -> BiMultimap a b

Construct a left-unique relation from a mapping from its right-elements to its left-elements.

Sets

dom :: BiMultimap a b -> Set a

Returns the domain of the relation, as a Set, in its entirety.

O(a).

ran :: BiMultimap a b -> Set b

Returns the range of the relation, as a Set, in its entirety.

O(a).

Relations

toRelation :: (Ord a, Ord b) => BiMultimap a b -> Relation a b

Convert a left-unique relation to a relation (forgetting its left-uniqueness).

Insert

insert :: (Ord a, Ord b) => a -> b -> BiMultimap a b -> BiMultimap a b

Insert a pair into a left-unique relation, maintaining left-uniqueness, preferring the latest inserted element.

That is, if a left-unique relation already contains the pair (x, y), then inserting the pair (z, y) will cause the (x, y) pair to be deleted.

unsafeInsert :: (Ord a, Ord b) => a -> b -> BiMultimap a b -> BiMultimap a b

Like insert x y, but the caller is responsible maintaining left-uniqueness.

Union

unsafeUnion :: (Ord a, Ord b) => BiMultimap a b -> BiMultimap a b -> BiMultimap a b

Union two left-unique relations together.

The caller is responsible for maintaining left-uniqueness.